pdf-to-excel-backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Google Cloud Services
google-cloud-storage==2.10.0
google-cloud-documentai==2.20.1
google-cloud-firestore==2.13.1

# PDF Processing
pdfminer.six==20221105
PyPDF2==3.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Word Document Processing
python-docx==1.1.0

# Firebase Authentication
firebase-admin==6.2.0

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
